Revert "unstable: set minimum ruby version to 2.4"
This reverts commit 5a62d5e7a23e01e5056fc6b61f589573a8641245.

We can't jump already into 2.4 because of JRuby.
It will support Ruby 2.4 as of 9.2.0.0.
If we bump Rubocop's `TargetRubyVersion`, it will make our builds to fail, due to the 2.4 specific suggestions.

On the other hand, if we bump already the min version in our gems 2.4,
the builds will still fail because JRuby 9.1 corresponds to MRI 2.3.

So we're locked for now.